Working multiple-store composition with querying.
The new MultiDatabaseTest shows how to use CompositeStore and the new LocatableSelectorBuilder
to create a storage architecture that decides where to store (and retrieve) things based on the
archetype details.

LocatableSelectorBuilder provides a convenient way to create customized rules for stores (and
parsers, and other locatable-handling things) to declare what they can and cannot handle.

Combining it with the CompositeStore allows presenting a homogenic store interface so clients
need not be aware of where data is stored.