Working & mostly tested /v2/locatable API.
Got rid of the unit test terminology and measurement service and 'fake' codes. When
using those from integration tests, they provide awkward results which don't work
with a real server. It's better to have realistic data there.

Messed around in medrecord-pv to support transforming not to a fully parsed Locatable
but just to a Map. This is then used for PATCH /v2/locatable/{id} (which still needs
tests).